backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for AI-generated content detection (text, images, videos) with chat and database features"
requires-python = ">=3.10"
dependencies = [
    # FastAPI core
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # ML dependencies
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "pytorchvideo>=0.1.5",
    "opencv-python>=4.8.0",
    "av>=10.0.0",
    "numpy>=1.24.0",
    "Pillow>=9.5.0",
    # SlowFast dependencies
    "matplotlib>=3.7.0",
    "tensorboard>=2.13.0",
    "pyyaml>=6.0",
    "yacs>=0.1.8",
    "simplejson>=3.19.0",
    "psutil>=5.9.0",
    "fvcore>=0.1.5",
    "iopath>=0.1.9",
    "fairscale>=0.4.13",
    "timm>=0.9.10",
    # ClipBased dependencies
    "open-clip-torch>=2.20.0",
    "huggingface-hub>=0.17.0",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "requests>=2.31.0",
    # Additional utilities
    "python-slugify>=8.0.0",
    "python-magic>=0.4.27",
    "sqlalchemy>=2.0.43",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "redis>=6.4.0",
    "aiocache>=0.12.3",
    "google-generativeai>=0.8.5",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.3.0",
    "python-socketio>=5.13.0",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "pytest-cov>=4.1.0"
]

# [project.scripts]
# Entry points should reference Python functions, not shell commands
# Use `uv run uvicorn main:app --host 0.0.0.0 --port 4001 --reload` directly

[tool.setuptools.packages.find]
include = ["slowfast_detection*", "clipbased_detection*", "api*", "core*", "schemas*", "services*", "utils*"]
exclude = ["tmp*", "tests*", "scripts*"]

[tool.uv]
# Outside Docker `uv sync` will default to .venv; inside Docker we use /opt/venv

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py310"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "ruff>=0.12.8",
]
